Move data to seed script and add db management fns
Since we only want to load our dummy data in development, the dummy
data is no longer loaded as part of a migration. It is now in a
separate seed sql file. You can load the data by running the
following:

  lein db populate

Also, added another helper that combines recreating the database and
populating it with data:

  lein db rebuild

Since we don't necessarily want to reset our database each time we
start the application in development, there is now a seperate setup
script that will take care of that. This is following the advice
from https://github.com/github/scripts-to-rule-them-all .