Thread database component to routes
Routes will now have access to the database component by using:

  (get-in request [:components :db])

I'm not pleased with this solution. Probably the right thing to do
here is to turn the routes into components. I've seen a number of
people do it that way to varying degrees of badly. Perhaps we will
revisit this once we get the basics of the app working.