Add logback and address java logging nightmare
Exceptions were happening on certain requests but nothing was being
logged, hence it is time to get the logging unbusted. There are two
popular log handlers in Javaland: log4j and logback. Both are made by
the same person, and you should use logback now (without getting into
the details, it's the author's recommendation).

To make matters more complicated, there are also two logging APIs in
Java: slf4j and commons-logging. Logback requires slf4j, but a lot of
libraries still use commons-logging. The lazy thing to do here is just
use and adapter so that log4j can be used with slf4j. This will work,
but that's living in the past. Logback, for a lot of reasons, is a
significant improvement to log4j.

The way we fix this problem is by including jcl-over-slf4j which is a
reimplementation of commons-logging that uses slf4j. Now our
dependencies that rely commons-logging can use slf4j. Next, we include
an empty commons-logging artifact to ensure commons-logging doesn't
get on the classpath and cause mischief. Then we can include
logback-classic. See, not so bad really.