Remove use of kosmos init-app
The use of init-app by the jetty component is problematic as it
mutates the var with our ring handler. Aside from philisophical issue
of mutating global state, this also breaks the ring reload middleware.
To resolve the issue, we instead have kosmos store the system in its
own global var and then use custom middleware to get the components we
need.

We came to kosmos for dependency injection, but ended up with
something more like a service locator. We can fix this, but not enough
time at the moment.