Add bootstrap
It's time to start making the UI not awful. There isn't a real great
way to include dependencies such as this in a project. Dependencies
are included via referencing the bootstrap CDN in the the layout.
Other options considered and their issues:

- bower/npm - requires yet another pair of tools that need to be
  installed in order to build project. Also, we'll still have to
  deal with Google Closure compiler issues with the JS.
- CLJSJS - great for JavaScript dependencies, sucks for things like
  CSS. If we used this, we'd still have to get the bootstrap CSS and
  other assets from somewhere else. Also, we'd then have to deal with
  ensuring we have the right assets for the right version of the JS.
- add files to project - this would be a quick fix to get out of this
  problem, but not hugely beneficial over pulling the stuff in from
  the CDN. Also, I dislike having a dependency's source code in the
  same repo as my own. There is a risk someone will modify the source
  without anyone else noticing.

We will, at some point, likely pull the JS in from CLJSJS, but we will
probably continue to get the other assets from the CDN.