Add ClojureScript support update/bootstrap scripts
ClojureScript added to project along with barebones namespace just to
see if things are building and loading correctly. This paves the way
for adding Om to the frontend.

Since we now have to compile the ClojureScript inorder for the
application to work, the update and bootstrap scripts have been added.
These scripts will take care of running database migrations and
ensuring the ClojureScript is compiled. The setup script has also been
modified to rebuild ClojureScript sources. Details in the README.md.